requirements.txt
streamlit
sqlalchemy
psycopg2-binary
python-dotenv
pandas
matplotlib
fpdf2
passlib
